Prep input for boardBits
