+++ includes/boardBits.hpp
@@ -1,0 +1,9 @@
+#ifndef h_BOARDBITS_h
+#define h_BOARDBITS_h
+
+#include "cells.hpp"
+
+
+
+
+#endif
