+++ includes/gol.hpp
@@ -1,41 +1,9 @@
+#ifndef h_GOL_h
+#define h_GOL_h
+
+#include "cells.hpp"
+
 #include <vector>
-
-class Cell
-{
-public:
-    Cell( bool live = false ) 
-        : live( live )
-        , nextLive( false ) 
-        , checked( false )
-        {};
-    ~Cell(){};
-    
-    void update()
-    {
-        live = nextLive;
-        nextLive = false;
-        checked = false;
-    }
-
-    bool checked;
-    bool live;
-    bool nextLive;
-};
-
-class CellPos
-{
-public:
-    CellPos( int posX = 0, int posY = 0, int index = 0 )
-    : posX( posX )
-    , posY( posY )
-    , index( index )
-    {};
-    ~CellPos(){};
-
-    int posX;
-    int posY;
-    int index;
-};
 
 class GolBoard
 {
@@ -49,11 +17,12 @@ private:
 public:
     void readInput();
     void updateBoard();
-    void updateCells();
     void printOutput();
 
 private:
     int getIndex( int x, int y );
     void checkNeighbors( std::vector< CellPos > pos, bool enable );
-
+    void updateCells();
 };
+
+#endif
